+++ app/bundles/PageBundle/Config/config.php
@@ -220,13 +220,7 @@ return [
             'mautic.page.subscriber.redirect' => [
                 'class'     => \Mautic\PageBundle\EventListener\RedirectSubscriber::class,
                 'arguments' => [
-                    'request_stack',
-                    'mautic.helper.ip_lookup',
-                    'mautic.lead.model.lead',
-                    'mautic.page.model.page',
-                    'monolog.logger.mautic',
-                    'mautic.lead.helper.primary_company',
-                    'translator',
+                   'mautic.page.helper.redirect',
                 ],
             ],
         ],
@@ -388,6 +382,19 @@ return [
                     'mautic.helper.core_parameters',
                     'request_stack',
                     'mautic.tracker.contact',
+                ],
+            ],
+            'mautic.page.helper.redirect' => [
+                'class'     => \Mautic\PageBundle\Helper\RedirectHelper::class,
+                'arguments' => [
+                    'request_stack',
+                    'mautic.helper.ip_lookup',
+                    'mautic.lead.model.lead',
+                    'mautic.page.model.page',
+                    'monolog.logger.mautic',
+                    'mautic.lead.helper.primary_company',
+                    'translator',
+                    'mautic.page.model.redirect',
                 ],
             ],
         ],
